Move view grid to the insight directory
