+++ client/web/src/enterprise/insights/components/insights-view-grid/SmartInsightsViewGrid.tsx
@@ -5,11 +5,11 @@ import { Layout, Layouts } from 'react-grid-layout'
 
 import { TelemetryProps } from '@sourcegraph/shared/src/telemetry/telemetryService'
 
-import { ViewGrid } from '../../../../views'
 import { Insight } from '../../core'
 import { getTrackingTypeByInsightType } from '../../pings'
 
 import { SmartInsight } from './components/SmartInsight'
+import { ViewGrid } from './components/view-grid/ViewGrid'
 import { insightLayoutGenerator, recalculateGridLayout } from './utils/grid-layout-generator'
 
 interface SmartInsightsViewGridProps extends TelemetryProps {
