+++ client/web/src/enterprise/insights/components/insights-view-grid/components/view-grid/ViewGrid.module.scss
