+++ client/web/src/enterprise/insights/components/insights-view-grid/components/view-grid/ViewGrid.story.tsx
@@ -3,11 +3,11 @@ import { noop } from 'lodash'
 import DotsVerticalIcon from 'mdi-react/DotsVerticalIcon'
 import FilterOutlineIcon from 'mdi-react/FilterOutlineIcon'
 
-import { Button, Menu, MenuButton, MenuItem, MenuList, Position } from '@sourcegraph/wildcard'
+import { Button, Menu, MenuButton, MenuItem, MenuList, Position } from '@sourcegraph/wildcard/out/src'
 
-import { WebStory } from '../../../components/WebStory'
-import { LINE_CHART_CONTENT_MOCK } from '../../mocks/charts-content'
-import * as View from '../view'
+import { WebStory } from '../../../../../../components/WebStory'
+import * as View from '../../../../../../views/components/view'
+import { LINE_CHART_CONTENT_MOCK } from '../../../../../../views/mocks/charts-content'
 
 import { ViewGrid } from './ViewGrid'
